third_party/blink/web_tests/flag-specific/highdpi/editing/selection/offset-from-point-expected.txt

12 345 6789

12 345 6789

12 345 6789

12 345 6789

LTR
0=-1 (-1)
1=11 (12)
2=31 (20)
3=51 (20)
4=71 (20)
5=91 (20)
6=111 (20)
7=131 (20)
8=151 (20)
9=171 (20)
10=191 (20)
11=211 (20)
LTR (Complex Path)
0=-1 (-1)
1=11 (12)
2=31 (20)
3=51 (20)
4=71 (20)
5=91 (20)
6=111 (20)
7=131 (20)
8=151 (20)
9=171 (20)
10=191 (20)
11=211 (20)
RTL
11=-1 (-1)
8=11 (12)
9=31 (20)
10=51 (20)
7=71 (20)
6=91 (20)
4=111 (20)
5=131 (20)
3=151 (20)
2=171 (20)
1=191 (20)
0=211 (20)
RTL (Complex Path)
11=-1 (-1)
8=11 (12)
9=31 (20)
10=51 (20)
7=71 (20)
6=91 (20)
4=111 (20)
5=131 (20)
3=151 (20)
2=171 (20)
1=191 (20)
0=211 (20)
VERTICAL-LR
0=-1 (-1)
1=11 (12)
2=31 (20)
3=51 (20)
4=71 (20)
5=91 (20)
6=111 (20)
7=131 (20)
8=151 (20)
9=171 (20)
10=191 (20)
11=211 (20)
VERTICAL-LR (Complex Path)
0=-1 (-1)
1=11 (12)
2=31 (20)
3=51 (20)
4=71 (20)
5=91 (20)
6=111 (20)
7=131 (20)
8=151 (20)
9=171 (20)
10=191 (20)
11=211 (20)
VERTICAL-RL
0=-1 (-1)
1=11 (12)
2=31 (20)
3=51 (20)
4=71 (20)
5=91 (20)
6=111 (20)
7=131 (20)
8=151 (20)
9=171 (20)
10=191 (20)
11=211 (20)
VERTICAL-RL (Complex Path)
0=-1 (-1)
1=11 (12)
2=31 (20)
3=51 (20)
4=71 (20)
5=91 (20)
6=111 (20)
7=131 (20)
8=151 (20)
9=171 (20)
10=191 (20)
11=211 (20)
